Update some code and add more test for BE

+++ coffee_shop_management_backend/module/ingredient/ingredientmodel/filter.go
@@ -4,7 +4,7 @@ type Filter struct {
 	SearchKey   string   `json:"searchKey,omitempty" form:"search"`
 	MinPrice    *float32 `json:"minPrice,omitempty" form:"minPrice"`
 	MaxPrice    *float32 `json:"maxPrice,omitempty" form:"maxPrice"`
-	MinAmount   *float32 `json:"minAmount,omitempty" form:"minAmount"`
-	MaxAmount   *float32 `json:"maxAmount,omitempty" form:"maxAmount"`
+	MinAmount   *int     `json:"minAmount,omitempty" form:"minAmount"`
+	MaxAmount   *int     `json:"maxAmount,omitempty" form:"maxAmount"`
 	MeasureType string   `json:"measureType,omitempty" form:"measureType"`
 }
